benchmarks/requirements.txt
faker>=22.0.0
pydantic>=2.5.0
playwright>=1.40.0
google-generativeai>=0.3.0
pdf2image>=1.16.0
pillow>=10.0.0
python-dotenv>=1.0.0
